pyproject.toml
[project]
name = "browser-use"
description = "Make websites accessible for AI agents"
authors = [{ name = "Gregor Zunic" }]
version = "0.1.40"
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx>=0.27.2",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "posthog>=3.7.0",
    "playwright>=1.51.0",
    "markdownify==1.1.0",
    "langchain-core==0.3.35",
    "langchain-openai==0.3.1",
    "langchain-anthropic==0.3.3",
    "langchain-ollama==0.2.2",
    "google-api-core>=2.24.0",
    "pyperclip>=1.9.0",   # only used for examples
]
urls = { "Repository" = "https://github.com/browser-use/browser-use" }

[project.optional-dependencies]
dev = [
    "tokencost>=0.1.16",
    "build>=1.2.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "fastapi>=0.115.8",
    "inngest>=0.4.19",
    "uvicorn>=0.34.0",
    "langchain>=0.3.18",
    "langchain-aws>=0.2.11",
    "langchain-fireworks>=0.2.6",
    "langchain-google-genai==2.0.8",
]

[tool.ruff]
line-length = 130
select = ["E", "F", "I"]
fix = true

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "browser_use/**/*.py",
  "!browser_use/**/tests/*.py",
  "!browser_use/**/tests.py",
  "browser_use/agent/system_prompt.md",
  "browser_use/dom/buildDomTree.js",
]

